# Don't deepcopy AST when generating vars