@amitkanfer 2024-11-21
> @swiatekm I think we want to backport this to 8.17 also. WDYT?

Are we 100% sure there is no hidden bug here?