@pierrehilbert 2024-11-21
My two cents: 
- We don't have the first BC for 8.17 yet
- We are backporting this to 8.15 and 8.16 so I hope we don't have an hidden bug here as the timeframe to test for those will be shorter than the 8.17 branch.