@swiatekm 2024-11-21
If anything, maybe we should skip the 8.15 backport? It's pretty close, and this isn't that critical of a fix. For 8.17, I think we should backport either way, but I could wait until 8.17.0 is out before merging the PR.

> > @swiatekm I think we want to backport this to 8.17 also. WDYT?
> 
> Are we 100% sure there is no hidden bug here?

I'm never going to claim 100% for a codebase I haven't been working on for at least a year. I am as sure as I can reasonably be, I think. In practice, the change only has an effect when one of the dynamic variable providers is used. For these:

* The local dynamic provider is used in unit tests
* I've done fairly extensive manual e2e testing on the kubernetes provider

I could also manually test the docker provider, but I can't think of any reason it would be affected when the others aren't.